+++ src/ShowSlider.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include <qboxlayout.h>    // for QHBoxLayout
+#include <qpointer.h>      // for QPointer
 #include <qslider.h>       // for QSlider
 #include <qspinbox.h>      // for QSpinBox
 #include <qstring.h>       // for QString
 #include <qtmetamacros.h>  // for Q_OBJECT
 #include <qwidget.h>       // for QWidget
-#include <QPointer>
 
 class ShowSlider : public QWidget {
   Q_OBJECT
@@ -17,8 +17,8 @@ class ShowSlider : public QWidget {
   const QPointer<QSpinBox> spin_box_pointer = new QSpinBox();
   const QPointer<QSlider> slider_pointer = new QSlider();
   const QPointer<QHBoxLayout> layout_pointer = new QHBoxLayout();
-  
-  explicit ShowSlider(int minimum, int maximum, const QString& suffix,
-             QWidget* parent = nullptr);
-  void set_value_override(int new_value);
+
+  explicit ShowSlider(int minimum, int maximum, const QString &suffix,
+                      QWidget *parent = nullptr);
+  void set_value_override(double new_value);
 };
